[DEBT-043-C] Split api/client.py into endpoint modules
- Split 1724-line client.py into _mixins/ package with 11 files, all ≤300 lines
- Created _base.py (134 lines) with ClientBase infrastructure (HTTP, rate limiting, retry)
- Created endpoint mixins: markets.py (297), events.py (223), series.py (73),
  exchange.py (32), multivariate.py (58), portfolio.py (122), trading.py (286),
  orders.py (273), order_groups.py (195)
- Thin client.py (144 lines) composes classes via mixin inheritance
- KalshiClient inherits from KalshiPublicClient (backward-compatible)
- All 200+ API unit tests pass
- Acceptance criteria: 3/4 complete (Phase C done, Phase D remaining)
